+++ staramr/subcommand/Search.py
@@ -55,11 +55,15 @@ class Search(SubCommand):
                                                 help='Search for AMR genes')
 
         self._default_database_dir = AMRDatabasesManager.get_default_database_directory()
+        default_database_repos = AMRDatabasesManager.create_default_manager().get_database_repos()
+
         cpu_count = multiprocessing.cpu_count()
 
         arg_parser.add_argument('--pointfinder-organism', action='store', dest='pointfinder_organism', type=str,
-                                help='The organism to use for pointfinder {' + ', '.join(
-                                    PointfinderBlastDatabase.get_available_organisms()) + '}. Defaults to disabling search for point mutations. [None].',
+                                help=(f'The organism to use for pointfinder. '
+                                      f"Validated: {set(default_database_repos.get_valid_pointfinder_organisms())}. "
+                                      f"All: {set(default_database_repos.get_pointfinder_organisms())}. "
+                                      f"Defaults to disabling search for point mutations. [None]."),
                                 default=None, required=False)
         arg_parser.add_argument('--plasmidfinder-database-type', action='store', dest='plasmidfinder_database_type',
                                 type=str,
@@ -293,10 +297,20 @@ class Search(SubCommand):
 
             logger.info("Finished. Took %s minutes.", time_difference_minutes)
 
+            included_pointfinder = pointfinder_database is not None
+
             settings = database_repos.info()
 
             settings['mlst_version'] = JobHandler.get_mlst_version(JobHandler)
             settings['command_line'] = ' '.join(sys.argv)
+            settings['pointfinder_organism'] = pointfinder_database.organism if included_pointfinder else 'None'
+
+            if included_pointfinder and not pointfinder_database.is_validated():
+                settings['messages'] = (f'Warning: Selected organism [{pointfinder_database.organism}] is '
+                                        f'not part of the validated set of organisms for PointFinder:'
+                                        f' {set(pointfinder_database.get_available_organisms())}. Cannot guarantee that all '
+                                        f'point mutations were detected properly.')
+
             settings['version'] = self._version
             settings['start_time'] = start_time.strftime(self.TIME_FORMAT)
             settings['end_time'] = end_time.strftime(self.TIME_FORMAT)
@@ -352,9 +366,15 @@ class Search(SubCommand):
 
         resfinder_database = database_repos.build_blast_database('resfinder')
         if (args.pointfinder_organism):
-            if args.pointfinder_organism not in PointfinderBlastDatabase.get_available_organisms():
-                raise CommandParseException("The only Pointfinder organism(s) currently supported are " + str(
-                    PointfinderBlastDatabase.get_available_organisms()), self._root_arg_parser)
+            if args.pointfinder_organism not in database_repos.get_pointfinder_organisms():
+                raise CommandParseException(f"The organism \"{args.pointfinder_organism}\" is not found in the selected PointFinder database. The "
+                                            f"only organisms available are: {set(database_repos.get_pointfinder_organisms())}. "
+                                            f"Of these, only {set(PointfinderBlastDatabase.get_available_organisms())} have been validated.",
+                                            self._root_arg_parser)
+            elif args.pointfinder_organism not in PointfinderBlastDatabase.get_available_organisms():
+                logger.warning("The only validated Pointfinder organism(s) are " + str(
+                    set(PointfinderBlastDatabase.get_available_organisms())) + f'. By selecting "{args.pointfinder_organism}" you are not guaranteed ' \
+                    + 'that all point mutations in this PointFinder database will be properly detected.')
             pointfinder_database = database_repos.build_blast_database('pointfinder',
                                                                        {'organism': args.pointfinder_organism})
         else: